Removed Keras directory. That should NOT have been committed in the first place.
Renamed the linked list file to SinglyLinkedList.py This is because that's the only type of linked list in that file. There is now a separate file for Doubly Linked list. I haven't fully implemented the double linked list class yet, I am just committing the structure of it while cleaning up the rest of the files.

Added the solution for HackerRank problem to reverse a linked list.