Some Updates to the PdfSplitTool.
I shrunk the scope of the general try catch to the actual problem area. The "problem" is a permissions issue based on how virus scanner software is working on the user's computer.
Shrinking the scope of that try catch will make it so that if a file can't be renamed, the whole program will just continue to the next file instead of quiting out.

I also added a count of files that were split and files that already were just one page.